control/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
grpcio==1.69.0
grpcio-tools==1.69.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kafka-python-ng==2.2.3
MarkupSafe==3.0.2
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.1
protobuf==5.29.3
requests==2.32.3
setuptools==75.8.0
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
etcd3==0.12.0
openai==1.63.2